moves duplicated method to user class
